backend/requirements.txt
fastapi>=0.104.1,<0.105.0
uvicorn>=0.24.0,<0.25.0
python-multipart>=0.0.6
google-generativeai>=0.7.1
google-ai-generativelanguage>=0.6.10
pydantic>=2.6.4,<3.0.0
starlette>=0.27.0,<0.28.0  # Aligned with FastAPI 0.104.1 requirements
python-dotenv>=1.0.0
typing-extensions>=4.8.0
webrtcvad
numpy
python-magic